# GPTutor-Backend-v2/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GPTutor Backend v2 API
  version: 2.0.0
  description: |
    # GPTutor API

    GPTutor предоставляет единый API для доступа к различным AI-моделям через один endpoint. Автоматически обрабатывает fallback'и и выбирает наиболее экономичные варианты.

    ## Быстрый старт

    ### Использование с OpenAI SDK

    **Python:**
    ```python
    from openai import OpenAI

    client = OpenAI(
        base_url="http://localhost:3001/v1",
        api_key="your-api-key-here"
    )

    completion = client.chat.completions.create(
        model="google/gemini-2.5-flash-lite",
        messages=[
            {"role": "user", "content": "Привет! Как дела?"}
        ]
    )

    print(completion.choices[0].message.content)
    ```

    **JavaScript:**
    ```javascript
    import OpenAI from 'openai';

    const openai = new OpenAI({
      baseURL: 'http://localhost:3001/v1',
      apiKey: 'your-api-key-here'
    });

    const completion = await openai.chat.completions.create({
      model: 'google/gemini-2.5-flash-lite',
      messages: [
        { role: 'user', content: 'Привет! Как дела?' }
      ]
    });

    console.log(completion.choices[0].message.content);
    ```

    ### Прямое использование API

    ```bash
    curl -X POST http://localhost:3001/v1/chat/completions \
      -H "Authorization: Bearer your-api-key-here" \
      -H "Content-Type: application/json" \
      -d '{
        "model": "google/gemini-2.5-flash-lite",
        "messages": [
...

...

...

...
servers:
...
paths:
  /v1/chat/completions:
    post:
      tags: [Chat Completions]
      summary: Создание chat completion
      description: |
        Создание chat completion с поддержкой streaming.
        
        ## Параметры
        
        - `model` - модель для генерации (по умолчанию: google/gemini-2.5-flash-lite)
        - `messages` - массив сообщений для контекста
        - `stream` - включить streaming режим
        - `max_tokens` - максимальное количество токенов
        - `temperature` - креативность ответа (0.0-1.0)
        
        ## Streaming
        
        При `stream: true` ответ приходит в формате Server-Sent Events:
        
        ```
        data: {"choices":[{"delta":{"content":"Hello"}}]}
        data: {"choices":[{"delta":{"content":" there!"}}]}
        data: [DONE]
        ```
        
        ## Стоимость
        
        Стоимость рассчитывается автоматически и списывается с баланса пользователя. Информация о стоимости включена в каждый ответ.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatCompletionRequest'
            examples:
              simple:
                summary: Простой запрос
                value:
                  model: "google/gemini-2.5-flash-lite"
                  messages:
                    - role: "user"
                      content: "Привет! Как дела?"
              streaming:
                summary: Streaming запрос
                value:
                  model: "google/gemini-2.5-flash-lite"
                  messages:
                    - role: "user"
                      content: "Расскажи историю"
                  stream: true
                  max_tokens: 150
              advanced:
                summary: Продвинутый запрос
                value:
                  model: "anthropic/claude-3-haiku"
                  messages:
                    - role: "user"
                      content: "Объясни квантовую физику простыми словами"
                  max_tokens: 500
                  temperature: 0.7
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatCompletionResponse'
              examples:
                non_streaming:
                  summary: Обычный ответ
                  value:
                    id: "chatcmpl-abc123"
                    object: "chat.completion"
                    created: 1677652288
                    model: "google/gemini-2.5-flash-lite"
                    choices:
                      - index: 0
                        message:
                          role: "assistant"
                          content: "Привет! У меня всё отлично, спасибо!"
                        finish_reason: "stop"
                    usage:
                      prompt_tokens: 4
                      completion_tokens: 147
                      total_tokens: 151
                      cost: 0.0000592
                      prompt_tokens_details:
                        cached_tokens: 0
                        audio_tokens: 0
                      cost_details:
                        upstream_inference_completions_cost: 0.0053
                      completion_tokens_details:
                        reasoning_tokens: 0
                        image_tokens: 0
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events для streaming режима
              examples:
                streaming:
                  summary: Streaming ответ
                  value: |
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1677652288,"model":"google/gemini-2.5-flash-lite","choices":[{"index":0,"delta":{"role":"assistant","content":""},"finish_reason":null}]}
                    
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1677652288,"model":"google/gemini-2.5-flash-lite","choices":[{"index":0,"delta":{"content":"Привет"},"finish_reason":null}]}
                    
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1677652288,"model":"google/gemini-2.5-flash-lite","choices":[{"index":0,"delta":{"content":"!"},"finish_reason":null}]}
                    
                    data: {"id":"chatcmpl-abc123","object":"chat.completion.chunk","created":1677652288,"model":"google/gemini-2.5-flash-lite","choices":[{"index":0,"delta":{},"finish_reason":"stop"}]}
                    
                    data: [DONE]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '402':
          $ref: '#/components/responses/InsufficientBalance'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    ApiKeyBearer:
      type: http
      scheme: bearer
      description: |
        API ключ пользователя для доступа к LLM endpoints.
    VkAuth:
      type: http
      scheme: bearer
      description: |
        VK подпись для авторизации через VK приложение.
        Передается в заголовке Authorization как Bearer token.
  schemas:
    ChatCompletionRequest:
      type: object
      required: [messages]
      properties:
        model:
          type: string
          default: "google/gemini-2.5-flash-lite"
          example: "google/gemini-2.5-flash-lite"
          description: |
            Модель для генерации ответа. Поддерживаются все модели OpenRouter.
            
            **Популярные модели:**
            - `google/gemini-2.5-flash-lite` - быстрая и дешевая
            - `anthropic/claude-3-haiku` - сбалансированная
            - `openai/gpt-4o-mini` - OpenAI модель
        messages:
          type: array
          items:
            $ref: '#/components/schemas/ChatMessage'
          example:
            - role: "user"
              content: "Привет!"
        max_tokens:
          type: integer
          minimum: 1
          maximum: 4000
          example: 150
          description: Максимальное количество токенов в ответе
        temperature:
          type: number
          minimum: 0
          maximum: 2
          example: 0.7
          description: Креативность ответа (0 = детерминированный, 2 = очень креативный)
        top_p:
          type: number
          minimum: 0
          maximum: 1
          example: 0.9
          description: Nucleus sampling параметр
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
          example: 0
          description: Штраф за повторение слов
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
          example: 0
          description: Штраф за повторение тем
        stop:
          type: array
          items:
            type: string
          example: ["\n", "###"]
          description: Стоп-последовательности для завершения генерации
        stream:
          type: boolean
          default: false
          example: true
          description: Включить streaming режим (Server-Sent Events)
    ChatMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [system, user, assistant]
          example: "user"
          description: |
            Роль отправителя сообщения:
            - `system` - системные инструкции для AI
            - `user` - сообщения пользователя  
            - `assistant` - ответы AI
        content:
          type: string
          example: "Привет! Как дела?"
          description: Текст сообщения
    ChatCompletionResponse:
      type: object
      properties:
        id:
          type: string
          example: "chatcmpl-abc123"
        object:
          type: string
          example: "chat.completion"
        created:
          type: integer
          example: 1677652288
        model:
          type: string
          example: "google/gemini-2.5-flash-lite"
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChatChoice'
        usage:
          $ref: '#/components/schemas/Usage'
    ChatChoice:
      type: object
      properties:
        index:
          type: integer
          example: 0
        message:
          $ref: '#/components/schemas/ChatMessage'
        finish_reason:
          type: string
          enum: [stop, length, content_filter]
          example: "stop"
    Usage:
      type: object
      description: |
        Информация об использовании токенов и стоимости.
        
        **Структура ответа:**
        - `prompt_tokens` - токены запроса
        - `completion_tokens` - токены ответа  
        - `total_tokens` - всего токенов
        - `cost` - цена в долларах
        - `cost_details.upstream_inference_completions_cost` - цена в рублях
        - `prompt_tokens_details` - детали токенов запроса
        - `completion_tokens_details` - детали токенов ответа
      properties:
        prompt_tokens:
          type: integer
          example: 4
          description: Количество токенов в запросе
        completion_tokens:
          type: integer
          example: 147
          description: Количество токенов в ответе
        total_tokens:
          type: integer
          example: 151
          description: Общее количество токенов
        cost:
          type: number
          format: float
          example: 0.0000592
          description: Стоимость запроса в долларах
        prompt_tokens_details:
          type: object
          properties:
            cached_tokens:
              type: integer
              example: 0
              description: Количество кешированных токенов
            audio_tokens:
              type: integer
              example: 0
              description: Количество аудио токенов
        cost_details:
          type: object
          properties:
            upstream_inference_completions_cost:
              type: number
              format: float
              example: 0.0053
              description: Стоимость запроса в рублях
        completion_tokens_details:
          type: object
          properties:
            reasoning_tokens:
              type: integer
              example: 0
              description: Количество токенов рассуждения
            image_tokens:
              type: integer
              example: 0
              description: Количество токенов изображений
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid API key"
...
  responses:
    UnauthorizedError:
      description: Ошибка авторизации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            invalid_api_key:
              summary: Неверный API ключ
              value:
                error: "Invalid API key or inactive user"
            missing_api_key:
              summary: Отсутствует API ключ
              value:
                error: "Missing or invalid Authorization header"
...
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missing_messages:
              summary: Отсутствует массив messages
              value:
                error: "messages array is required"
    InsufficientBalance:
      description: Недостаточно средств
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            low_balance:
              summary: Недостаточно средств
              value:
                error: "Insufficient balance"
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            server_error:
              summary: Ошибка сервера
              value:
                error: "Internal server error"